split up bars index
