+++ src/traces/bars/arrays_to_calcdata.js
@@ -1,0 +1,26 @@
+/**
+* Copyright 2012-2015, Plotly, Inc.
+* All rights reserved.
+*
+* This source code is licensed under the MIT license found in the
+* LICENSE file in the root directory of this source tree.
+*/
+
+
+'use strict';
+
+var mergeArray = require('../../lib').mergeArray;
+
+
+// arrayOk attributes, merge them into calcdata array
+module.exports = function arraysToCalcdata(cd) {
+    var trace = cd[0].trace,
+        marker = trace.marker,
+        markerLine = marker.line;
+
+    mergeArray(trace.text, cd, 'tx');
+    mergeArray(marker.opacity, cd, 'mo');
+    mergeArray(marker.color, cd, 'mc');
+    mergeArray(markerLine.color, cd, 'mlc');
+    mergeArray(markerLine.width, cd, 'mlw');
+};
